Fixing unused variable warnings.
r5397

+++ framework/src/bcs/FunctionPeriodicBoundary.C
@@ -24,7 +24,7 @@ FunctionPeriodicBoundary::FunctionPeriodicBoundary(SubProblem & subproblem, std:
     _tr_y(fn_names.size() > 1 ? &subproblem.getFunction(fn_names[1]) : NULL),
     _tr_z(fn_names.size() > 2 ? &subproblem.getFunction(fn_names[2]) : NULL)
 {
-  if (_dim != subproblem.mesh().dimension())
+  if (_dim != static_cast<int>(subproblem.mesh().dimension()))
     mooseError("Transform function has to have the same dimension as the problem being solved.");
 }
